head_model/pyproject.toml
[project]
name = "head_model"
version = "0.1.0"
description = "Training scripts and tools for the Beaker bird head detection model"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core training and preprocessing dependencies
    "numpy>=2.3.2",
    "opencv-python>=4.8.0",
    "ultralytics>=8.3.172",
    "platformdirs>=4.0.0",
    "pillow>=11.3.0",
    "polars>=1.32.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "ruff>=0.7.0",
    "pip-audit>=2.7.0",
    "pre-commit>=3.8.0",
    "comet-ml>=3.41.0",
    "onnx>=1.12.0,<1.18.0",
    "onnxslim>=0.1.59",
    "onnxruntime>=1.22.0",

]

[project.scripts]
# Training and data processing scripts
beaker-train = "train:main"
beaker-convert = "convert_to_yolo:main"
beaker-upload = "upload_train:main"

[project.urls]
"Homepage" = "https://github.com/ericphanson/beaker"
"Repository" = "https://github.com/ericphanson/beaker"
"Documentation" = "https://github.com/ericphanson/beaker#beaker-training-tools"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "E722", # bare except (sometimes needed for broad exception handling)
    "F841", # assigned but never used (common in tests)
    "B007", # loop variable not used in body (common in enumerate)
    "E402", # module level import not at top (needed for dynamic imports)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
